requirements.txt
omxplayer-wrapper
python-osc
